state -> renderer and introduce demand renderer
